+++ src/main/java/bossmonster/Main.java
@@ -4,6 +4,7 @@ import java.util.Scanner;
 
 import bossmonster.controller.Controller;
 import bossmonster.domain.monster.BossMonsterDamageGenerator;
+import bossmonster.service.Service;
 
 public class Main {
     public static void main(String[] args) {
