+++ src/main/java/bossmonster/domain/characterattributes/Hp.java
@@ -1,4 +1,4 @@
-package bossmonster.domain.info;
+package bossmonster.domain.characterattributes;
 
 public class Hp {
