+++ src/main/java/bossmonster/service/Service.java
@@ -1,11 +1,11 @@
-package bossmonster;
+package bossmonster.service;
 
 import java.util.List;
 
-import bossmonster.domain.info.Hp;
-import bossmonster.domain.info.Mp;
-import bossmonster.domain.info.Name;
-import bossmonster.domain.info.Stat;
+import bossmonster.domain.characterattributes.Hp;
+import bossmonster.domain.characterattributes.Mp;
+import bossmonster.domain.characterattributes.Name;
+import bossmonster.domain.characterattributes.Stat;
 import bossmonster.domain.monster.BossMonster;
 import bossmonster.domain.player.Attack;
 import bossmonster.domain.player.Player;
